requirements.txt
asgiref==3.8.1
aspose-cells-python==24.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
cssbeautifier==1.14.7
Django==4.1.13
django-annoying==0.10.6
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-pandas==0.6.6
django-pyodbc==1.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djhtml==2.0.0
djlint==1.19.14
EditorConfig==0.12.3
et-xmlfile==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
imap-tools==1.0.0
IMAPClient==2.3.1
jsbeautifier==1.14.7
Markdown==3.4.1
mssql-django==1.2
mysqlclient==2.2.6
numpy==2.0.0
openpyxl==3.1.2
pandas==2.2.2
paramiko==3.0.0
pathspec==0.11.0
pillow==10.2.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyodbc==4.0.35
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.14
wfastcgi==3.0.0
xmltodict==0.13.0
